+++ DatadogTrace/Sources/Span/SpanEventBuilder.swift
@@ -10,13 +10,15 @@ import DatadogInternal
 /// Builds `SpanEvent` representation (for later serialization) from span information recorded in `DDSpan` and values received from global configuration.
 internal struct SpanEventBuilder {
     /// Service name to encode in span.
-    let serviceName: String?
+    let service: String?
     /// Enriches traces with network connection info.
     /// This means: reachability status, connection type, mobile carrier name and many more will be added to every span and span logs.
     /// For full list of network info attributes see `NetworkConnectionInfo` and `CarrierInfo`.
     let networkInfoEnabled: Bool
     /// Span events mapper configured by the user, `nil` if not set.
     let eventsMapper: SpanEventMapper?
+    /// If spans should be enriched with the current RUM context.
+    let bundleWithRUM: Bool
     /// Telemetry interface.
     let telemetry: Telemetry
 
@@ -45,6 +47,20 @@ internal struct SpanEventBuilder {
         let regularTags = castValuesToString(tagsReducer.reducedSpanTags)
         tags.merge(regularTags) { _, regularTag in regularTag }
 
+        if bundleWithRUM {
+            // Enrich with RUM context
+            do {
+                if let rum: RUMContext = try context.baggages[RUMContext.key]?.decode() {
+                    tags[SpanTags.rumApplicationID] = rum.applicationID
+                    tags[SpanTags.rumSessionID] = rum.sessionID
+                    tags[SpanTags.rumViewID] = rum.viewID
+                    tags[SpanTags.rumActionID] = rum.userActionID
+                }
+            } catch let error {
+                telemetry.error("Failed to decode RUM context for enriching span", error: error)
+            }
+        }
+
         // Transform user info to `SpanEvent.UserInfo` representation
         let spanUserInfo = SpanEvent.UserInfo(
             id: context.userInfo?.id,
@@ -58,7 +74,7 @@ internal struct SpanEventBuilder {
             spanID: spanID,
             parentID: parentSpanID,
             operationName: operationName,
-            serviceName: serviceName ?? context.service,
+            serviceName: service ?? context.service,
             resource: tagsReducer.extractedResourceName ?? operationName,
             startTime: startTime.addingTimeInterval(context.serverTimeOffset),
             duration: finishTime.timeIntervalSince(startTime),